backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
scikit-learn
pymc
arviz
numpy
pandas
pickle